requirements.txt
flask==3.0.0
tensorflow==2.15.0
numpy==1.24.3
Pillow==10.0.0
opencv-python==4.8.1.78
werkzeug==3.0.1
matplotlib==3.7.2